chore[]: adiciona roteamento de stack
