+++ app/_layout.tsx
@@ -1,10 +1,10 @@
 import { Stack } from "expo-router";
 
 export default function RootLayout() {
-  return <Stack
-    screenOptions={{
-      headerShown: false,
-    }}>
-      <Stack.Screen name="index" />
-</Stack>;
+  return ( 
+    <Stack>
+        <Stack.Screen name="index" />
+        <Stack.Screen name="supplierRegistration" options={{headerTitle: "Registrar Novo Produto"}}/>
+    </Stack>
+  );
 }
